👌🚚 rename utils to serializationUtils
