+++ packages/rum-recorder/src/domain/rrweb-snapshot/index.ts
@@ -1,5 +1,5 @@
 import { serializeNodeWithId, transformAttribute, IGNORED_NODE, snapshot } from './snapshot'
 export * from './types'
-export * from './utils'
+export * from './serializationUtils'
 
 export { snapshot, serializeNodeWithId, transformAttribute, IGNORED_NODE }
