+++ packages/rum-recorder/src/domain/rrweb-snapshot/serializationUtils.ts
